feat(docs): Add initial README.md and project structure  Added initial contents to the README.md. Note that the README.md is not complete yet as the project is still in progress. I will update it in the final commit to ensure it is comprehensive and purposeful.  Added a public folder with the following structure: - css/style.css - images/ (for images to be included in the README.md) - js/ (for JavaScript logic for the static page)  Added a views folder with the following structure: - views/home.handlebars - views/dash.handlebars - views/layouts/main.handlebars - views/partials/summary.handlebars
feat(docs): Add initial README.md and project structure

Added initial contents to the README.md. Note that the README.md is not complete yet as the project is still in progress. I will update it in the final commit to ensure it is comprehensive and purposeful.

Added a public folder with the following structure:
- css/style.css
- images/ (for images to be included in the README.md)
- js/ (for JavaScript logic for the static page)

Added a views folder with the following structure:
- views/home.handlebars
- views/dash.handlebars
- views/layouts/main.handlebars
- views/partials/summary.handlebars