feat(server): add configuration and form flip animation  Added a connection.js to include a configuration for the server. Included the server modules in the server.js file. Included the necessary packages in the package.json. Included styles in the styles.css file located in the public folder nested in the css folder for styling the login and the signup page to have a flip animation to switch between the login and signup forms. Added package-lock.json to the .gitignore. Added a signup-login.js file in the public nested folder "js" to handle flipping the form. Added html to the home.handlebars for the form.
feat(server): add configuration and form flip animation

Added a connection.js to include a configuration for the server. Included the server modules in the server.js file. Included the necessary packages in the package.json. Included styles in the styles.css file located in the public folder nested in the css folder for styling the login and the signup page to have a flip animation to switch between the login and signup forms. Added package-lock.json to the .gitignore. Added a signup-login.js file in the public nested folder "js" to handle flipping the form. Added html to the home.handlebars for the form.